✨ feat: connect payment data right away
